amd64: add a basic ACPI emulation to enable SMP
In order to enable x86 SMP support for UVMM, we need to emulate enough
of ACPI so that the guest OS finds the local APICs in the Multiple APIC
Description Table (MADT) and uses them to send startup IPIs from the
Bootstrap Processor (BSP) to the Application Processors (AP).

This implementation writes the minimum number of ACPI tables into guest
memory so that Linux successfully discovers the APs enabled in the
system. It therefore has a number of limitations, mainly that we only
provide an empty DSDT:

The Differentiated System Description Table (DSDT), as described in the
ACPI Spec v6.3 describes a system's peripherals. The description is done
using Definition Blocks encoded in AML, which are rather complex. Other
Open Source Hypervisors have addressed this problem in a number of ways:

 * Qemu builds the DSDT using a custom AML builder.
 * Xen uses a tool called mk_dsdt to generate some parts of the .asl
   source files from templates and iasl -tc to generate a C char array
   representation of the AML code.
 * VirtualBox uses a static DSDT and uses indexed registers to control
   the presence of certain devices.

As a consequence of the lack of device descriptions in the DSDT, we
currently cannot enable Hardware-reduced ACPI to deprecate legacy parts
of ACPI that we don't implement, because with this flag switched on in
the FADT, a Linux 4.19 guest will rely on the DSDT to find PCI devices.

Fixes #CD-24.

Change-Id: Ie0629b3c7a146b24f433bdbc8c8f925152c83425