Fix optee smccc func check
The check was too generic and did cover other service call ranges
(including the once for PSCI) as well.

This can be problematic when optee uses HVC and is registered as the
first service. It will than accept all the PSCI calls, but obviously
will not do the right thing.

Now we check for the trusted OS/application range explicitly (the 32/64
bit check is done in the generic handler already).

Fixes #CD-76

Change-Id: Ieb167ebd2eb7f57ae5052176283dbeedbe5e29d5