show node name on fatal device creation failure
Device creation can fail moderately when the create_dev() just returns a
nullptr. In that case the error message often contains the corresponding
device name. Or an exception is thrown in which case there is often not
enough information about which device is affected.

In the former case the device creation just fails but the VM starts up
anyway (without the failing device). In the latter case the VM start
fails fatal. For now we don't change this behavior.

Change-Id: I9b77c777d7b75b38eb5335bc90476da26a8bf657