Pass vcpu pointer to Pt_walker::walk
Future implementations of page table walkers for architectures other
than AMD64 will require access to a multitude of system registers via a
vcpu pointer passed to them. This change anticipates this by changing
the interface of the AMD64 implementation of Pt_walker::walk such that
it can easily be generalized across architectures.

Change-Id: I8591b51596be666bd127ce6e1a7f81317d03ca5c