PCI: rework of the MSI memory management
The page containing the MSI memory is acquired by a Ds_access_mgr.
This manager interfaces with MMIO areas on top of it via the
Mmio_ds_converter, which contains a pointer to the manager and
the offset to memory inside the managed region.
The Mmio_ds_converter can act as standalone read/write pass-through
device or act as an interface from another object to the
Ds_access_mgr.

The Virt_msix_table serves guest read/writes with local memory and only
on a MSI-X entry unmask configures the routing from the device to the
VMM and guest.

Change-Id: Iae282142d5ba110b8ffd421fd56ba5d7dc2595de