amd64: enable 2 CPUs in device tree and set maximum CPUs to 8
This commit finishes the x86 SMP work by enabling 2 CPUs in our base
device tree and setting the maximum supported CPUs to 8. The later limit
corrensponds to the current default for ARM and was chosen because some
resources are allocated per supported CPU, independent of the number of
CPUs enabled at run time.

Fixes #CD-23.

Change-Id: I21051fb11d3dbf43110f3f244854100e10d42d00