Add remap support to Ds_handler
Make it possible to remap mmio regions in the Ds_handler to a new
address. Those regions are not mergeable and therefore have to be marked
with the new moveable flag.

Also add a Map_eager flag which is on by default as before but make it
possible to leave it out in case this is not a desired behavior.

Change-Id: I15f3ab7753f4f21e5584b0463852a6692f981de1