amd64: handle non-IRR interrupts in the local APIC
Up to now, the PIC sent a normal IRQ to the IOAPIC stub which just
forwarded the IRQ to the first local APIC. This interface does not
transport enough information.

The delivery mode for instance determines if the IRQ goes into the
IRR or is handled outside of the IRR queue.
The local APIC now offers a new interface to handle a MSI data type
as interrupt information. All APIC clients handling MSIs use this new
interface.

The PIC sends its interrupts in the MSI address/data format
with delivery mode ExtINT, which is not queued in the local APIC's IRR.

With this commit SMI, NMI, INIT, and ExtINT interrupts are handled
outside of the IRR.

Change-Id: Ied5c8a134bc04229e8de7b7a11bc114e13c65b8c