amd64: Handle MMIO access to APIC more thoroughly
When accessing local APIC's ICR, care must be taken to differentiate the
access method as there are subtle differences when accessing ICR via an
MSR, in which case ICR behaves as a 64-bit register, or as an MMIO
register, in which case ICR behaves as two consecutive 64-bit registers
that each hold the respective 32-bits in its lower part.

Since uvmm internally translates MMIO local APIC accesses to MSR
accesses, the MSR read/write functions are extended to take a third
argument, which indicates the actual access method.

Fixes #CD-151.

Change-Id: I9414a87bad476c4a42cb3e3560fc588d2bb0da77