amd64: set up guest CPUs in prepare_platform()
For amd64, the vCPUs and a number of CPU-related virtual devices such as
the local APIC have historically been set up in run().

This commit introduces the new Guest method prepare_platform() to
separate the setup of virtual devices that exist in Guest and are not
part of the device tree.

Change-Id: I62c1488c421762f5b97b0978b920fa38b9c27584