amd64: handle Inter-Processor Interrupts (IPIs)
On x86, Inter-Processor Interrupts are sent by writing to the Interrupt
Command Register (ICR). Besides IPIs, special INIT and Startup delivery
modes are used to start Application Processors (APs) from the bootstrap
processor (BSP).

This change introduces an ICR handler that has access to the CPU
container and the MSI-X controller. To avoid circular dependencies, the
local APIC access handler passes ICR accesses to a separate class before
forwarding other accesses to the local APIC. Depending on the delivery
mode, the ICR handler advances the startup sequence of an AP or forwards
the IPI message to the MSI-X Controller for distribution to a remote
processor's local APIC.

Fixes #CD-156.

Change-Id: I904e265f9736078f9e3485f3d30028ec91d82baf