amd64: run one clock thread per physical core
Because we set up VMX to virtualize the RDTSC instruction, it is
necessary to start one timer thread per vCPU and reference its local
time stamp counter, as otherwise the TSC value read from the clock would
not correspond to the value returned by the RDTSC instruction.
The legacy PIT continues to run off the timer thread on the guest's BSP.
Local APIC timers now register on their vCPU's timer thread.
In case of the KVM clock, according to Documentation/virtual/kvm/msr.txt
in the Linux source tree, the tsc_timestamp field represents "the tsc
value at the current vCPU at the time of the update of this structure."
It therefore also registers at the clock thread for its vCPU.

Rescheduling a guest CPU's timer thread on the scheduler ID of its
physical core requires that the CPU information has already been parsed
from uvmm's Device Tree configuration. However, the legacy PIT is also
specified via the Device Tree and at construction needs a clock source
to register with. We therefore create a static array of Clock_source
objects that timers can register with and defer the creation of timer
threads to the initialization of CPU structures in
Guest::prepare_platform(), which runs after the guest's CPU structures
are populated and scheduling information for each vCPU is available.

Fixes #CD-206.

Change-Id: Ib948132caa66ac0fc4773b823c0906ef9d78f857