separate arch specific code from MSI and PCI files
The code contains AMD64 specific code like the Lapic_array, src_id
computation, and Virt_msix_table creation. Architecture specifics must
not be in generic code.

The decoding of the MSI is done in the architecture specific MSI routing
in the Msix_controller. This controller must be known to the PCI bus,
therefore the controller is added in the device tree file.

The translation from the IRQ object to the guest MSI is done in Msi_src,
which sends the MSI address, data pair to the Msix_controller, when
triggered.

The MSI formats move from msi.h to msi-arch.h.

For pci_bus_bridge.cc to remain architecture independent, the creation
of the Virt_table_msix moves to architecture specific code as well.
Furthermore, the register_msix_table_page and register_msix_bar function
are made static and removed from class Pci_bus_bridge.

Change-Id: Id3755dca8525d56d03c862268fcb8883fdb70723