ensure MSI writes reached the hardware
We must ensure writes of the device memory reached the device,
before we service reads from the emulated memory area.
The only relevant field is the MSI vector control, as it allows
masking and unmasking of the MSI. Only on reads to this field, we
read from the device memory.

Change-Id: Ia867c2b1edd1690e9101f34ed5c1f979ceb91fee