detect ELF binaries without PT_LOAD sections
If an ELF binary does not have a PT_LOAD section then the image load
range start/end values are not set and later there is an error message
that address 0xffffffff has no corresponding entry in guest2host().
For convenience, we add a proper error message for this case.

Change-Id: I0399557f6cd27e9b089396864c61fb09c06607f4