amd64: lock Feature Control MSR
The Feature Control Architecturial MSR (3AH, IA32_FEATURE_CONTROL
according to Vol. 4, 2.1 of the Intel SDM) can be used to enable
advanced processor features related to Intel Trusted Execution
Technology. As we currently only filter specific CPUID feature flags,
the availability of such features in the host processor may be visible
to the guest.

This commit sets the lock bit to signal the guest that writes to this
MSR are not supported. At this point, all other bit field flags are
unset, i.e. none of the features controlled by this MSR are enabled for
the guest. An attempt to enable any of the flags despite the presence of
the lock will result in an unsupported MSR write and halt execution in
uvmm, which corresponds well to the expected hardware behavior of
causing a General Protection Fault.

Change-Id: Ib9aabe376ff8de0d7137996403af5ed6bf2f957f