amd64: disable CPUID bits for AMD speculation control
For fine-grained speculation control, qemu 4.1.1 enables AMD CPUID bits
even when when runing with -cpu host on an Intel CPU. When passed on to
Linux running under uvmm, those CPUID leafs lead to an unsupported write
to the Speculation Control Architecturial MSR (48H, IA32_SPEC_CTRL
according to Vol. 4, 2.1 of the Intel 64 and IA-32 Architectures
Software Developer's Manual) during AP initialization. With the Linux
Spectre V2 mitigations introduced in commit 7b62ef142c11 and released as
part of v4.19.7 of the Linux kernel, this change is also needed on a
uniprocessor configuration.

Currently, uvmm uses blacklisting instead of whitelisting of CPUID bits,
so this change just removes the AMD bits for speculation control when
virtualizing the CPUID instruction.

Change-Id: I302e0e1cbca32632a8cf5790b78074ddea4f1a8f