extend Message Data Register width to 64bit
This corresponds to the Intel SDM Vol. 3A 10-35 cited in the original
implementation. Of the users of the send method, those sending PCI table
entries with 32bit data fields as specified in the PCI 3.0 specification
in Figure 6-11: MSI-X Table Structure will just be upcasted to the 64bit
unsigned values expected by the MSI-X controller.

With the Data_register_format structure extended to 64bit, upon
construction from a 64bit unsigned integer, the reserved upper 32bit can
be filled with data. This data is not accessible via bitfield members,
just as it was previously possible to fill reserved areas in the lower
32bit that whould only be accessible through the structure's raw()
method.

Change-Id: I6623026b5bc699cce002eff147ab505575b94da2