Unify initialization of Irq_svr, fix check for IO irqs
Irq_svr is used in three different places, in Io_proxy, Optee and
Pci_host_ecam_generic. In each case we use the same code sequence
to initialize the Irq-svr instance. Since all three instances throw
a runtime error if there is an error in the initialization we move
the setup code to the constructor of Irq_svr, which either succeeds
or throws a runtime error.

It may happen that an irq provided by IO is assigned multiple times
to the same or different devices. The Io_proxy tries to handle this
case by checking whether an already bound irq is identical to the one
it tries to bind. It relies an the fact, that in an earlier version it
was the only user of Irq_svr. So it could check whether the bound IRQ
is managed by Irq_svr and is bound to the same IO irq. That changed when
others started to use Irq_svr to manage Irqs provided by other services
than IO. Here we introduce a derived class Io_irq_svr to ensure that the
test works as it is supposed to do.

Change-Id: Ic398c499158c15b66e9cabcc6dd0bf3761255985