Add notion of a 'main vCPU'
This changes add the concept of a main vCPU to
Generic_cpu_dev. The 'main vCPU' is the vCPU stucture for the
main thread of the uvmm, and setup very early.
This allows us to use information of the supported features
of the microkernel's vCPU implementation already during
creation of the VM object and during device-tree scanning.

Change-Id: I1195fbf878fb01d8a7a74baf1d9236cecbcf08c0