amd64: send full ICR in IPI message on MMIO access
When accessing the Interrupt Command Register (ICR) via memory mapped
I/O, a write to the lower half causes the Inter-Processor Interrupt to
be sent. While setting up the ICR value correctly, our code incorrectly
only supplied the just written (lower) half in the IPI message. This led
to missing bits in the destination field, which in turn caused an
attempt to execute the AP bringup sequence for the BSP (CPU 0).
The issue is only present on MMIO access; when the register is accessed
via the X2APIC MSR, the value is supplied as one 64bit value, as opposed
to the two 32bit writes via MMIO.

Change-Id: I7e10ba36746ed37e58f679198d7bd573944d76bd