ARM: cosmetic issues and readability
Follow-up to Change-Id:Ia4b8903e75e3fc046b984c19f510eed01ef32128.

The change in Dist::_demux_irq_block() should improve the readability.
The static assertions intend to show that this change is valid and that
the generated code doesn't change under the assumption that the lower
bits of the passed register offset are properly aligned.

Also s/0u/0U/ because the latter form is generally used in Uvmm.

Add comments in Dist::_read_gicd_header() to show which registers are
actually read.

Add a dedicated case for GICD_STATUS in Dist::dist_read() because that
register is readable and always returns 0. And use register offsets
instead of 'reg >> 4' values because that's easier to associate with
the hardware. The generated code is the same.

Change-Id: Ie36cdfcc84eaa0695570aaa345669eb9a3b0c982