x86: fix source ID computation for PCI-MSI devices
io expects a source ID to identify the device on the PCI bus. This
source ID is passed to io, when the MSI information to program into the
device are requested via msi_info().

Part of this source ID is the device-function number of the device. uvmm
manages this as two 16bit numbers is Devfn_address and we just binary
or'ed this number to the source ID marker of 0x4'0000.
Looking into io revealed that this violates the expected format as we
mix the dev-part of the number with the marker bit at position 18.
The format expected by io encodes the device-function number in the
lower 8 bits of the source ID as follows:
  [2:0] function number
  [7:3] device number

This commit adds a function to Devfn_address to compute this format.
Furthermore, Devfn_address is now a parameter of make_virt_msix_table().

Change-Id: I04824d146b3c177c26312495475ddee072adb9c1