Tiny ACPI support
This change introduces a little bit of Acpi platform
emulation. Enough to make Linux use Acpi shutdown to turn off
the machine.

This requires "pci=noacpi" on the linux commandline to make
Linux use PCI for device discovery.

Notice that this is enabled in the default device tree.
However, a user can disable it with no ill effects (e.g. the
system will still be able to discover its CPUs, but it will no
longer support system shutdown).

Change-Id: I7a1f199bd48ecf096073d0b7cee5a13b2c418a5c