amd64: initial support for ACPI suspend (S3)
On a suspend call, the VM suspends until any interrupt triggers a wake
up.

The guest stores the wakeup vector in the ACPI FACS table and triggers
the suspend action by writing the Slp_en and Slp_type bits to the
Pm1a_cmd_block port.
On resume, the wakeup vector in the FACS table serves as entry point
to start the core again.

The address of the FACS table in guest address space needs to be marked
as reserved in the guest's e820 map.

Change-Id: I0d2da08484f0ed58bcc06edc9326bb98a56f87fe