device_tree: pass string length to fdt_getprop()
The libfdt library provides fdt_getprop() and fdt_getprop_namelen(). A
call to

  fdt_getprop(fdt, offs, name, lenp);

is translated by the library into

  fdt_getprop_namelen(fdt, offs, name, strlen(name), lenp);

We shall call fdt_getprop_namelen() directly and pass the string length
as parameter. This allows the compiler to determine the string length at
compile time in certain cases saving a call to strlen().

The performance benefit is not huge but measurable for big device trees.

Change-Id: I03b2dc5c3b0d3e3f6fc3428e66f29d18f77021ef