setprop_string() write optimization
On setprop_string() we try first to replace the property value in place.
In that particular case we don't need to flush the cache because no node
offset changed.

If that doesn't work (either the property does not exist or the stored
property value has a different size than the new value), fall back to
the regular function for setting a string property.

Change-Id: Ifb4d4a48c8aeca0636a76d13aed8ef88557d3d87