add optimizing function to disable a node
Having a central function to disable a device tree node is more
convenient than setting the "status" property to "disabled".

Linux guests only considers "ok" and "okay" as valid devices (or nodes
without any status at all). Therefore it would be sufficient to set the
"status" property to "disa" to disable the device. If there is already a
status property which is set to "okay" then writing this property is
just an in-place replacement without changing any node offset. Hence we
wouldn't need to flush the device cache.

Unfortunately the documentation is not clear about disabled nodes so
play safe and write "disabled".

Change-Id: Ie01ccff4d010049f53c43a1d343daf74ac50b587