vmx: ensure that no interrupt is lost on injection
This change adds a check for events that are in the process of
being injected. Only if no injection is happening do we inject
an interrupt.

Previously it could happen that we inject an event (by writing
the fields in the vmcs) and try to do a world switch. Now if
that call to the microkernel returns to uvmm before switching
to the guest, we might have tried to inject a second
interrupt, overwriting the one that was "in flight".

Change-Id: I578b0828a13b6f3e353b50153c129b54b625cb31