optimize a few Node::get_path() ops away
During device tree initialization, there is a lookup in the existing
devices if the device has already been initialized. The function

  Device_repository::device_from_node()

checks if the node has a valid phandle. If the node doesn't have a
phandle, the full device tree path is determined using the expensive
Node::get_path() function. If the device is not part of the _devices
repository then the function

  Device_repository::add()

adds the corresponding node to the _devices repository. The function
writes the node phandle (could be invalid) and determines the full
device tree path again to store both values into the _devices
repository.

Now the path value is only determined once and the value is used for
adding the node to the _devices repository.

Change-Id: I0e66ebb0050831e687c9bc45f4135d30b7760d51