modules: Make module uncompression explicit
Up to now we have implicitly uncompressed every
module. This is counterproductive for cases where
the module shall be compressed, e.g. for a compressed
ramdisk/initrd which guests can unpack themselves.
For this reason, we only uncompress when requested
via the new 'uncompress' option.
Uncompressing is still useful, for example, to save
space on the local disk.

Change-Id: Id1455b1cce3533cbd187c253ac5f7e5281bb6a9b