Mark classes final
Marking a class final allows the compiler to do certain optimizations on
virtual function calls. If a class was marked final and a call to a
virtual method is done when a object of this final type is used the
compiler can optimize away the vtable lookup. But most importantly it
may inline that function. This will of course not work if a base class
pointer is used (as indented with virtual method calls). However, this
still might compensate for the vtable lookup penalty in some cases. The
final uvmm executable gets slightly bigger which suggests some functions
where inlined.

Change-Id: I405ccdd56ed04e66647b8db972e640a35584e872