host_dt: copy the Fdt content not the object
The Dtb::Fdt object has two constructors, one taking a void* and the
other taking a Fdt const & and a padding which defaults to 0.

Host_dt contains a member Fdt* and provided this member to the Fdt
ctors. The valid ctor is the void* ctor, which set the passed pointer
as void *_dtmem, leading to a new Fdt object pointing to the old
Fdt object, not the intended old _dtmem.

We actually want to pass the Fdt object to the ctor, not the pointer
to it.

Change-Id: Iab1ba48759c43fab90faeafa6ab500b609a94458