do setup_ramdisk() earlier
The setup_ramdisk() function loads the RAM disk from the corresponding
file into the guest memory. As this function does not depend on the vbus
capability from the IO server, execute this function before scanning the
device tree + initializing the IO devices to increase the parallelism
between Uvmm and IO during boot. The benefit grows with the size of the
RAM disk.

All information required for loading the RAM disk are already present
when the kernel was loaded and the command line was set up.

For example, on RCar H3 it takes up to 10ms to load a 5MB RAM disk.

Change-Id: Ie1ae34cbc7de7eb35b66149b34aa7e32d6362ea5