amd64: PT walker: Write all exception info into msg
The PT walker throws exceptions to halt the uvmm in case a page table
walk fails.
Right before throwing it prints the non-present entry.

This is intentional for the 'uvmm needs to do MMIO decoding'-case.
However, in case the guest debugger is used to dump guest memory, the
"VMM: FATAL:" output is wrong, as the exception is caught in the guest
debugger code.

Instead of Err().printf we use sprintf to create a formated string
which is then used in the exception.

Change-Id: I0fb9b6ec9ebe5a102ef37d1e05f308b1b4f0b39a