Rework VMM internal guest memory handling
This change splits code that handles guest mapping of memory and mmio
from code that handles VMM local mapping of memory and mmio. This allows
for cleaner interfaces and better and cleaner support of the MAP_OTHER
feature.

The change also avoids a couple of dynamic casts during emulation of
guest RAM accesses in the VMM by using the Vm_ram object instead of the
more complicated and general Vm_memmap.

NOTE: This change currently disables AMD64 PCI pass-thru because this is
depending on invalid assumptions according to real MMIO to guest MMIO
address mappings.

Change-Id: Ifcf1231987cad3225808e2c4be6d69aa49ccab6c