amd64: save FPU+SSE state of guest
The FPU, SSE and AVX state is not saved by Fiasco in the vCPU state.
This has to be done by uvmm itself. In an ideal world uvmm would not
even touch the SSE state but in practice gcc emits such code by default
and it is hard to prevent this for all involved libraries.

Currently gcc only emits SSE instructions by default. We use fxsave64
to save just this state. Even though the upper eight registers are not
yet used by gcc we cannot use fxsave because it will not save all bits
of the FPU IP and DP registers. Whenever gcc starts emitting AVX
instructions this will have to be extended to use xsave.

Change-Id: Id42cee16fc620c7d4f763ecd48638d004b226787