VIO proxy: announce only two MSIs
The current l4virtio implementation uses a single IRQ to notify the
driver about virtqueue handling. This was not displayed correctly to the
guest when the Virtio-PCI transport is used.
The number of MSIs the device can handle suggested one interrupt per VQ.

The interaction of virtio-net and the PCI version of the virtio-proxy
lead to the situation, where the guest thought it has per VQ interrupts,
but virtio-net send only a shared interrupt.
On interrupt, the driver only processed the first VQ, which in case of
virtio-net is the RX queue. Eventually, a lot of spurious interrupts led
linux to disable the interrupt.

When linux sends network packets, the driver fills TX queue, but is not
notified about processed elements in the used ring.
Therefore, the sending stalls.

The reduction of the available MSIs to two mimics the behavior of
l4virtio where one MSI is used for the configuration interrupt and
one for the shared virtqueue interrupt.

Fixes ticket #CD-260.

Change-Id: I52d0186a53d3f4a8120301edbc2bc0828b14667c