Use nullptr instead of &_addr as mmap addr
According to the documentation of mmap, the first argument shall be a
null pointer or an address (void *) to map to. In this case, erroneously
the pointer of _addr (void **) was provided. Even (the probably intended
_addr (instead of &_addr) would still be incorrect, because _addr is
initialised with MAP_FAILED which is not a valid value for the first
mmap argument.

Change-Id: I79a23d2b608add24f477febf8de249f9a170124c