Uvmm: fix BAR size detection in PCI emulation
When an OS determines the size of a PCI BAR, it writes a -1 to
the BAR and reads it again. The size can then be calculated
with a mask and a 1 complement. The previous implementation
gave the plain size value (without flags) that would give the
OS a wrong BAR size.

Change-Id: I86694ed90b7c500bb093f0070599c699b6097995