arm: make register access thumb mode compatible
The inline assembly assumes that it is compiled as arm code because it
computes jumps. This won't work if the compiler is emitting thumb code
because the offsets are different.

One option would be to force the functions to be emitted as arm code.
This would imply to prevent any inlining, though. Instead, we go for
the other route, that is to not rely on inline assembly. The generated
code does look not much worse (GCC12.2: an additional jump table on arm
mode, a TBB instruction in thumb mode) so there shouldn't be a
noticeable performance impact.

Change-Id: Ic2bf4ef866fa32165234ba172539d087850bf7aa