gic: insert all pending IRQs sorted by priority
The guest access to the ICV_PMR (virtual priority mask register) is not
trapped. It it thus not possible to (re-)insert lower priority
interrupts if the guest lowers its priority mask. Looking at the PMR
may delay interrupts (possibly forever) if the guest currently has set
a high priority mask at interrupt injection.

Looking at the guest PMR is unnecessary anyway as there should always
be the highest priority pending interrupts in the list registers. Only
the IRQ with the highest priority will ever be delivered to the guest
by the hardware.

Change-Id: I3386e1327da404e2490efd4d19de9b73e02f2172