arm: optimize pending IRQ lookup
Instead of searching the whole Irq_array for a pending interrupt we
maintain a list of pending&enabled interrupts for each CPU. Every time
an IRQ is made pending & enabled (p&e) it will be inserted into the
list of the destination CPU.

Because there can be only one destination CPU the GICv2 1:n model is
not fully supported anymore. Only the first online CPU in the target
list will receive an interrupt. This should pose no problem as our main
guest does not use this feature and 1:n interrupts are considered
deprecated anyways.

Special care must be taken to handle migration. To keep the
synchronization overhead low, only the target CPU is allowed to take
p&e interrupts from its list. If a p&e interrupt is re-targeted the
currently owning CPU must be woken up to take the interrupt out of its
list and push it to the right CPU. To keep this working even if CPUs
are offline the thread of vCPU 0 will take over this responsibility for
offline CPUs.

Change-Id: Icbfacbb7d82e47f94de17610ed739180ce3ab27c