Introduce per cpu registry
Properly dispatching IPCs on all vCPUs requires a registry per vCPU
thread. The registry for the main thread is statically allocated
because it is required very early in the initialization of uvmm. The
registries of additional CPUs are allocated on demand once the threads
are created. The per vCPU-registry is accessible through the
Generic_vcpu_ptr interface. This way other objects can bind their gates
to the desired vCPU if required.

Change-Id: If829ec3fbb4e270a4c5907f4210df679b0180101