gic: report supported security/routing model correctly
Our implementation does only support a single security state. Given
that we do not support EL3 in guests it is the only sensible option
anyway. This is correctly reported for GICv2 but on GICv3 the
GICD_CTLR.DS bit must be RAO/WI.

The routing model implementation of GICv3 is strictly using affinity
routing for SPIs. This must be reported in GICD_TYPER by the No1N bit.
Likewise, the GICD_ARE bit in GICD_CTLR needs to be RAO/WI too.
Previously bit 4 was incorrectly set in the redistributor GICR_CTLR
register as intended ARE-bit. This bit is actually RES0.

Change-Id: I5b3befee4394be0775ea150818e4a90dac7b0f1e
Requires-Change-Id: I80e81c35ee2212d716eeac15746f2f0a583060a8