PCI: parse device BARs dependent on header type
The structure of the PCI config space depends on the value in the
header type field of the config space.
A header type value of 0x00 defines the used header type to be of
type 0 for normal PCI devices.
A value of 0x01 or 0x02 defines the header type to be type 1 or
type 2 for bridge and cardbus devices, respectively.

This commit improves the initial BAR parsing to distinguish between
these two types. It refactors the access to the header type field as
well.
We do not support type 2 headers and throw an error if a device of
such a type is encountered.

Change-Id: I2865336c34c083f53ebb20d0f02a8c1b486e4149