introduce address space manager; change in defaults on ARM
The address space manager decides in which way the guests address space
shall be set up, to enable DMA access for all platforms with differing
HW features.

This commit changes the default behavior for RAM configuration on ARM
architecture. Up until this change the default behavior was identity
mapping from guest-physical memory to its corresponding host-physical
addresses.
Now, the default will map the RAM starting at 0x0 or a given RAM base,
unless and device tree and memory node is given. Then these address are
used.

Furthermore, the physmap flag in the device tree is deprecated.

The identity mapping behavior can be forced with the '-i' command line
flag.

There are five modes the ASM can operate in:
  No_dma, Dma_offset, Identity, Iommu and Iommu_identity.

If no vbus is given to uvmm, there is no need for DMA, thus, the No_dma
mode is selected.

If there is a vbus and we can assign a L4Re::Dma_space to it and there
is a 'dma-ranges' property in the device tree's memory node, the
Dma_offset mode is selected. In this mode, the guest is made aware of
the offset between it's RAM addresses and the DMA addresses via the
dma-ranges field in the DT's memory node.
If there is no 'dma-ranges' property, Identity mode is selected.

If there is a vbus and we force identity mode, the Identity mode is
selected. The guest RAM will be mapped to the corresponding
host-physical addresses of the RAM dataspace.

If there is a vbus and an IOMMU present, we select Iommu mode. The RAM
is mapped into the KDMA space under the addresses used by the guest.

If there is a vbus, an IOMMU and identity mode is forced, the guest RAM
will be mapped to the corresponding host-physical addresses of
the RAM dataspace. The KDMA space is then set up using the same
addresses.

If the ASM operates in 'No DMA' mode, even when there are DMA capabable
devices on the vBus, please check that the hardware device as the flag
'Io.Hw_device_DF_dma_supported'.

Change-Id: I9d41a6ca9fe80611b609347e999f9f46cca4397e