l4_cache_*() functions use the end address exclusive
The l4_cache_*() expect the start address of the region and the address
of the first byte after the region. Usually this is done by calling such
a function like

  l4_cache_clean_data(start, start + size);

This is already done on other places. Fix two places in virtio_dev.h
where the end address was specified as "inclusive".

Change-Id: I18e58c3efff4953c8b3b345b03c92d0bb76d79e0