pci: Add virtual source ID to PCI devices
MSIs on ARM with the GICv3 make it is necessary that each device that is
an MSI source has a device/source ID. Conceptually, this source ID is
transmitted as sideband data when the device performs a write access
that triggers an MSI, so that the GIC can securely identify the acting
device. Typically this source ID is also what an IOMMU would use to
identify devices.

We have to recreate this source ID mechanism for our virtualized GIC,
since the Interrupt Translation Service (ITS) component in the GIC,
responsible for converting MSIs to interrupts, mandates it. Our
virtualized ITS implements the Msix_controller interface, which
virtualized devices use to trigger MSIs. Previously, only the MSI
address and the MSI data were passed as arguments to
Msix_controller::send(). For the ITS, Msix_controller::send() has been
extended with an additional argument, the virtual source ID.

The Msix_controller implementation on x86 currently does not need the
virtual source ID, so it just ignores the additional argument.

A class named Msix_dest is introduced, comprised of a reference to an
Msix_controller and a virtual source ID. Msix_dest represents the
association of an MSI-X source, such as a PCI device that can send
MSI-X messages, to an MSI-X controller. In all places where previously
a reference to an Msix_controller was used to send an MSI-X message,
this has been replaced by an Msix_dest.

For PCI devices the source ID is usually derived from the PCI
configuration address (Bus/Device/Function) via the msi-map device tree
property(see linux/Documentation/devicetree/bindings/pci/pci-msi.txt).
Therefore, we cannot know the virtual source ID of a PCI device
until it is registered on the PCI host bridge, or more precisely, until
we know its PCI configuration address. Thanks to the introduction of a
device ID allocator in a previous change, this prerequisite is no
longer an issue.

Change-Id: I19132cf320a33c444591654e248d43b7f4814059