arm: Add MSI support for GICv3
In the GICv3 architecture the Interrupt Translation Service
(ITS) provides support for message-based interrupts, e.g. Message
Signaled Interrupts(MSI). The guest configures the ITS to map the
combination of a DeviceID and an EventID to a Locality-specific
Peripheral Interrupt (LPI) directed to a redistributor (i.e. to a CPU).
To trigger an LPI, a device has to write the corresponding EventID to a
register of the ITS. For our virtualized ITS, the devices perform this
write access via the Msix_controller interface.

To represent the state of LPIs the existing classes Irq_array and
Gic::Irq are used. The LPI Irq_array is managed by the Its class, which
provides a virtualized ITS to the guest. In addition, a pointer to the
LPI Irq_array is added to Gic::Cpu class, allowing irq_from_intid() to
lookup the LPI corresponding to an INTID. Furthermore, the Redist
class is extended with additional per vCPU state (Redist_cpu):
  - Control register (enable LPI support for redistributor)
  - LPI configuration table (LPI enable and priority)
  - LPI pending table (not used for now)

The GICv3 specification states that the ITS is an optional component. As
an alternative, GICv3 redistributors can support direct LPI injection
without an ITS as the middleman, i.e. the devices directly write into a
register of the redistributor. Although in theory we could use this as
shortcut to avoid implementing the rather complicated ITS component,
adding support for direct LPI injection in our virtualized
redistributors won't do us any good, because Linux does not support
LPIs without an ITS.

Change-Id: Iaa542b404ee8debc5e939e9a1403b801e6f97081