Add support for a simple framebuffer device
Here we add support for a simple framebuffer device as supported
by Linux (CONFIG_FB_SIMPLE=y, CONFIG_X86_SYSFB=y).

The device expects a node in the device tree as follows:
    simplefb {
        compatible = "simple-framebuffer";
        reg = <0x0 0xf0000000 0x0 0x1000000>;
        l4vmm,cap = "fb";
    };
The reg entry describes a window in the guest address space where
the framebuffer is placed. The l4vmm,cap property contains the name
of a cap providing a frame buffer via the goos protocol.

The emulation fetches the framebuffer from the server, fills the Linux
bootpage with parameters based on the provided framebuffer and maps
the framebuffer to the guest. There is not window/view handling or
refresh handling implemented (yet).

Change-Id: I690d00cf94e73257662be5429c85a3bab60cfc60