ARM: Change target architecture to armv7ve
Now that uvmm runs in PL1 mode, some features that are part of the ARMv7
virtualization extension (armv7ve) are available to uvmm, such as
access to banked registers via the mrs/msr instructions.

In order for the compiler to allow the use of this functionality the
target architecture has to be set to armv7ve. There is no point in
compiling uvmm for a non armv7ve architecture anyway, since Fiasco only
provides virtualization support on CPUs that implement the
virtualization extensions. The armv7ve target architecture is supported
by gcc since version 4.9.0.

Change-Id: Id8c6058f22ffbd957b22640b412e915fbe088991