ARM: Inject exceptions for unhandled faults
Until now, uvmm would simply stop the VM, if the guest triggers a fault
that uvmm cannot handle, for example accessing an invalid memory area.

However, this behavior is incomplete and also not very helpful when
troubleshooting in the guest. Therefore, in case of an unhandled fault,
uvmm now injects an exception into the guest instead of stopping the VM.
This allows the guest to either recover from the error or provide
helpful debug output. For example, a Linux guest kernel will report a
kernel panic.

For an unhandled memory fault uvmm injects a synchronous data or
prefetch abort into the guest. For an unknown fault uvmm injects an
undefined instruction exception.

When injecting an exception, the following three scenarios must be
considered separately, as the involved mechanisms and/or registers
differ:
  - Exception taken to Aarch64 mode on Aarch64 host
  - Exception taken to Aarch32 mode on Aarch64 host
  - Exception taken to Aarch32 mode on Aarch32 host

Aarch64@Aarch64 and Aarch32@Aarch32 are fundamentally different, but
both are unproblematic, since all required registers are accessible for
uvmm running in PL1. However, injecting an exception into an Aarch32
guest on an Aarch64 host would require extending Fiasco's extended vCPU
API, to make the SPSR_und, SPSR_abt and IFSR32_EL2 registers accessible
to uvmm. Apart from that, this scenario is already fully implemented,
partially reusing functionality from the Aarch32@Aarch32 scenario.

The Aarch32@Aarch64 scenario must not be confused with the case where
a fault that occurred in Aarch32 guest user mode is injected into an
Aarch64 guest kernel, which already works fine with the current state
of this change.

Change-Id: I3c7584a7f101c0c553d8aaf42ea91ca8f991cbf8